+++ app/components/Views/AddDeviceToWallet/showAlreadySyncedSheet.test.ts
@@ -1,6 +1,6 @@
-import { strings } from '../../../locales/i18n';
-import Routes from '../../constants/navigation/Routes';
-import type { AppNavigationProp } from '../NavigationService/types';
+import { strings } from '../../../../locales/i18n';
+import Routes from '../../../constants/navigation/Routes';
+import type { AppNavigationProp } from '../../../core/NavigationService/types';
 import { showAlreadySyncedSheet } from './showAlreadySyncedSheet';
 
 const mockNavigate = jest.fn();
