+++ app/components/Views/AddDeviceToWallet/showImportFailedSheet.test.ts
@@ -1,6 +1,6 @@
-import { strings } from '../../../locales/i18n';
-import Routes from '../../constants/navigation/Routes';
-import type { AppNavigationProp } from '../NavigationService/types';
+import { strings } from '../../../../locales/i18n';
+import Routes from '../../../constants/navigation/Routes';
+import type { AppNavigationProp } from '../../../core/NavigationService/types';
 import { showImportFailedSheet } from './showImportFailedSheet';
 
 const mockNavigate = jest.fn();
